+++ src/pages/index.jsx
@@ -4,7 +4,7 @@ import ContactForm from "../components/HomePage/ContactForm";
 import Services from "../components/HomePage/Services";
 import Landing from "../components/HomePage/Landing";
 import Work from "../components/HomePage/Works";
-import Footer from "../components/Footer";
+import Footer from "../components/Layouts/Footer";
 
 const IndexPage = () => {
 	return (
